+++ config/custom-environment-variables.json
@@ -50,7 +50,10 @@
     "endpoint": "S3_ENDPOINT",
     "accessKeyId": "S3_ACCESS_KEY_ID",
     "secretAccessKey": "S3_SECRET_ACCESS_KEY",
-    "sslEnabled": "S3_SSL_ENABLED",
+    "sslEnabled": {
+      "__name": "S3_SSL_ENABLED",
+      "__format": "boolean"
+    },
     "maxRetries": {
       "__name": "S3_MAX_RETRIES",
       "__format": "number"
