+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "service-name",
-  "version": "1.0.0",
-  "description": "This is template for map colonies typescript service",
+  "name": "exporter-cleanup",
+  "version": "2.0.0",
+  "description": "map exporter cleanup cron job",
   "main": "./src/index.ts",
   "scripts": {
     "test:unit": "jest --config=./tests/configurations/unit/jest.config.js",
