+++ src/common/enums.ts
@@ -2,3 +2,11 @@ export enum StorageProviderType {
   FS = 'FS',
   S3 = 'S3',
 }
+
+export enum OperationStatus {
+  PENDING = 'Pending',
+  IN_PROGRESS = 'In-Progress',
+  COMPLETED = 'Completed',
+  FAILED = 'Failed',
+  EXPIRED = 'Expired',
+}
